Adding comparison results enum
It's a bit clunky, but using booleans doesn't work for comparison well,
since equality isn't representable. As well integers are not perfect
since nobody really cares if the comparison results in -1 or -10, it's
always checked against 0. A strongly typed response gives the data
structures better information by restricting the results to useful
values.